Coerce candidates to string before highlighting
Some helm sources, such as the `helm-themes` package, use symbols in
their candidates list, which would break the assumption that they are
chars or stings somehwere in `helm-flx-fuzzy-highlight-match`.

This adds a coercion function `helm-flx-candidate-string` that converts
a symbol to a string -- with a cond statement, so this can be extended
to other types, though I'm not sure this would come up in practice? --
and ensures that candidates are strings in the scope of the highlighting
function.